% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_corporate_colors.R
\name{create_corporate_colors}
\alias{create_corporate_colors}
\title{Create a corporate color scheme}
\usage{
create_corporate_colors(factor_colors = c(red = "#d11141", green = "#00b260",
  blue = "#01afdc", yellow = "#f37735", orange = "#b301e3"))
}
\description{
This funciton will create a colorscheme for graphics, including tables and
plots as well as section headings
}
