% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_piclink.R
\name{create_piclink}
\alias{create_piclink}
\title{create a piclink object}
\usage{
create_piclink(links, name = links, pic = name, verbose = FALSE)
}
\arguments{
\item{links}{either a string, character vector, or named character vector
supplying the web address(es) to be linked}

\item{name}{the alt text and prefix of the image file name. suffixes will be
identified from the name and the pic assigned. defaults to the value of links}

\item{verbose}{I like to build in messages. Do you want to see them?}
}
\description{
a pic link object is a series of textless buttons for use in an HTML page,
but plans exist to modify this to accept a LaTeX list of pic links to align
out into a page for printed pdf use. the html generated here can also be used
in shiny apps by supplementing with the \code{HTML()} line.
}
