% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_sedna_object.R
\name{define_sedna_object}
\alias{define_sedna_object}
\title{define a sedna object}
\usage{
define_sedna_object(sedna_object_name = Sys.Date(),
  sedna_author = Sys.info()[["user"]],
  sedna_company = Sys.inf()[["machine"]], render_type = c("html", "latex",
  "shiny", "rmarkdown", "script"), style_template = c("cv", "resume",
  "startpage", "raw", "memo", "article"))
}
\arguments{
\item{sedna_object_name}{the official name of the object, will define to
whatever the object name is. Use this to customize the official reports of
title.}

\item{sedna_author}{the author you want displayed. will default to system user}

\item{render_type}{the type of object or objects to be rendered}

\item{style_template}{the style template you want to make.}

\item{sections}{a character vector of sections to be included. refer to details
for the specific default values for each type of render}
}
\description{
a sedna object contains all the information needed to comiple a report
these informations are all optional and can be used separately. some
objects must have certain attributes and may or may not break or complete
improperly without error.
}
\details{
this function called without a defined \code{sedna_object_name} will
generate all of the default possibilities. Providing details to this function
will really unlock a lot of the customization with the software.
}
