% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_chrome_history.R
\name{fetch_chrome_history}
\alias{fetch_chrome_history}
\title{fetch Chrome browser history}
\usage{
fetch_chrome_history()
}
\description{
this function will fetch all your Google Chrome browser history and store it
in a `chrome_history` psdlow style class with special methods in multiple
funcitons defined within this package. Currently, it only reads data from
<insert what it does here> while importing that data into a structured class.
}
