% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_img_pic.R
\name{find_img_pic}
\alias{find_img_pic}
\title{find an img pic in the image directory}
\usage{
find_img_pic(img, dir = "img/")
}
\arguments{
\item{img}{the possible name of the image}

\item{dir}{where the image file should be searched. defaults to img/}
}
\description{
this function finds a close match using grep to the files in a directory
}
