% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_startpage_files.R
\name{generate_startpage_files}
\alias{generate_startpage_files}
\title{Create HTML and CSS}
\usage{
generate_startpage_files(title = "my start page", layout_style = "simple",
  div_sections = c("main", "media", "shop", "social", "help", "fun", "sports",
  "lab", "learn"), main_links = c("gmail", "live", "github", "forbin",
  "rstudio", "trello"), media_links = c("plex@home", "plex@anywhere", "hbo",
  "plexpy", "netflix", "hulu", "flixster", "amazon", "spotify", "youtube",
  "soundcloud", "lma"), shop_links = c("amazon", "etsy", "craigslist", "ebay",
  "massdrop", "woot", "abe", "leftlane"), social_links = c("facebook",
  "linkedin", "reddit", "goodreads", "discogs", "boardgamegeek", "twitter",
  "pocket", "researchgate"), help_links = c("stackoverflow", "latex",
  "bioinformatics", "r-bloggers"), sports_links = c("espn", "covers",
  "volnation", "volquest", "tfp", "watchespn"), lab_links = c("molarity",
  "dilution", "solution", "moles", "concentration", "molar", "weigelworld",
  "mafft", "blast", "ncbi"), fun = c("dominion", "foe"))
}
\description{
This function is used to create the static HTML and CSS files to display a
customized startpage
}
