% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFigure.R
\name{generateFigure}
\alias{generateFigure}
\title{Generate a figure for inclusion in a LaTeX document}
\usage{
generateFigure(figure)
}
\arguments{
\item{figure}{a plot or \code{ggplot2} object, if it has the specified attributes}
}
\description{
This function writes a graphic to disk in the specified folder and returns
the LaTeX, markdown, or HTML code necessary to display it.
}
